Separate State Machine and Workflow concepts. Includes other refinements/restructuring for clarity.
